exception handling

+++ Ecommerce-backend/product/src/main/java/com/ecommerce/product/Service/impl/ProductServiceImpl.java
@@ -3,7 +3,11 @@ package com.ecommerce.product.Service.impl;
 import com.ecommerce.product.Entity.Product;
 import com.ecommerce.product.Repository.ProductRepository;
 import com.ecommerce.product.Service.ProductService;
+import com.ecommerce.product.exception.ProductAlreadyExistsException;
+import com.ecommerce.product.exception.ProductNotFoundException;
 import jakarta.transaction.Transactional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,57 +25,73 @@ import java.util.UUID;
 @Transactional
 public class ProductServiceImpl implements ProductService {
 
+    private static final Logger logger = LoggerFactory.getLogger(ProductServiceImpl.class);
+
     @Autowired
     private ProductRepository productRepository;
 
     @Override
     public Page<Product> getProducts(Pageable pageable) {
+        logger.debug("Fetching products with pageable: {}", pageable);
         return productRepository.findAll(pageable);
     }
 
     @Override
     public Product getProductById(String Id) {
-        Optional<Product> optionalProduct = productRepository.findById(Id);
-        return optionalProduct.orElse(null);
+        logger.debug("Fetching product with ID: {}", Id);
+        Product product = productRepository.findById(Id).orElseThrow(() -> new ProductNotFoundException("Product with ID " + Id + " not found"));
+        logger.debug("Product fetched: {}", product);
+        return product;
     }
 
     @Override
     public Product getProductByProductName(String productName){
+        logger.debug("Fetching product with productName: {}", productName);
         Optional<Product> optionalProduct = productRepository.findByProductName(productName);
-//        return optionalProduct.orElse(null);
-        if(optionalProduct.isEmpty())
-            return null;
-        return optionalProduct.get();
+        return optionalProduct.orElse(null);
     }
 
     @Override
     public Product createProduct(Product product) {
+        logger.debug("Creating product: {}", product);
         if(product.getProductId() == null)
             product.setProductId(UUID.randomUUID().toString());
         else {
             Optional<Product> existingProduct = productRepository.findById(product.getProductId());
             if (existingProduct.isPresent()) {
-                throw new RuntimeException("Product with ID " + product.getProductId() + " already exists.");
+                logger.error("Product with ID {} already exists.", product.getProductId());
+                throw new ProductAlreadyExistsException("Product with ID " + product.getProductId() + " already exists.");
             }
         }
+        logger.debug("Product created: {}", product);
         return productRepository.saveAndFlush(product);
     }
 
 
     @Override
     public void deleteProduct(String Id) {
+        if(!productRepository.existsById(Id)) {
+            logger.error("Product with ID {} not found.", Id);
+            throw new ProductNotFoundException("Product with ID " + Id + " not found.");
+        }
+        logger.debug("Deleting product with ID: {}", Id);
         productRepository.deleteById(Id);
     }
 
     @Override
     public Product updateProduct(Product updatedProduct) {
-        Product product = productRepository.saveAndFlush(updatedProduct);
-        return product;
+        if(!productRepository.existsById(updatedProduct.getProductId())) {
+            logger.error("Product with ID {} not found.", updatedProduct.getProductId());
+            throw new ProductNotFoundException("Product with ID " + updatedProduct.getProductId() + " not found.");
+        }
+        logger.debug("Updating product: {}", updatedProduct);
+        return productRepository.saveAndFlush(updatedProduct);
     }
 
     @Override
     public void patchProduct(String productId, Map<String, Object> updates) {
-        Product existingProduct = productRepository.findById(productId).orElseThrow(() -> new RuntimeException("Product not found"));
+        logger.debug("Patching product with ID: {} with updates: {}", productId, updates);
+        Product existingProduct = productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("Product with ID " + productId + " not found."));
 
         updates.forEach((key, value) -> {
             try {
@@ -99,5 +119,16 @@ public class ProductServiceImpl implements ProductService {
 
         productRepository.save(existingProduct);
     }
+
+//    @Override
+//    public byte[] export(String fileType) {
+//        logger.debug("Exporting products with file type: {}", fileType);
+//        List<Product> products = productRepository.findAll();
+//        List<ProductExportBean> productExportBeans = products.stream().map(product -> {
+//            ProductToProductExportBeanMapper mapper = new ProductToProductExportBeanMapper();
+//            return mapper.ProductToProductExportBean(product);
+//        }).toList();
+//        return PdfExport.exportProductData(productExportBeans);
+//    }
 
 }
